Consolidate AdminSet and AdminSetBehavior
There's no need for the to be a Class + Module as that pattern is only
needed when the class is generated into the application and intended to
be customized. That is not the case with AdminSet.
